pyproject.toml
[tool.poetry]
name = "flowerbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mubble = "^1.1.0"
tortoise-orm = "^0.21.3"
asyncpg = "^0.29.0"
aioowm = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
